+++ src/main/java/xyz/blackmonster/clients/BitCoinClient.java
@@ -2,5 +2,5 @@ package xyz.blackmonster.clients;
 
 public interface BitCoinClient {
 
-	float readPriceEUR();
+	double readPriceEUR();
 }
